refactor(vim): clean up plugin configuration
