+++ config/nvim/lua/plugins/nvimtree.lua
