+++ config/nvim/lua/plugins/trouble.lua
